Feat: add -d or --dry-run switch to output to console (#42)
* add -d or --dry-run switch to output to console

* bump version to v1.0.9

* update changelog

+++ src/command.js
@@ -19,8 +19,19 @@ class GenEnvTemplateCommand extends Command {
     },
   ]
 
+  static description = packageInfo.description
+
+  static flags = {
+    // add --version flag to show CLI version
+    version: flags.version({ char: 'v' }),
+    // add --help flag to show CLI version
+    help: flags.help({ char: 'h' }),
+    // add dry-run flag to output to the console
+    'dry-run': flags.boolean({ char: 'd' }),
+  }
+
   async run() {
-    const { args } = this.parse(GenEnvTemplateCommand)
+    const { args, flags: appFlags } = this.parse(GenEnvTemplateCommand)
 
     const envFileString = await promises.readFile(
       args.source,
@@ -28,17 +39,14 @@ class GenEnvTemplateCommand extends Command {
     )
 
     const templateFileString = genEnvTemplate(envFileString)
-    await promises.writeFile(args.dest, templateFileString, { encoding: 'utf-8' })
+
+    if (appFlags['dry-run']) {
+      // eslint-disable-next-line no-console
+      console.log(templateFileString)
+    } else {
+      await promises.writeFile(args.dest, templateFileString, { encoding: 'utf-8' })
+    }
   }
-}
-
-GenEnvTemplateCommand.description = packageInfo.description
-
-GenEnvTemplateCommand.flags = {
-  // add --version flag to show CLI version
-  version: flags.version({ char: 'v' }),
-  // add --help flag to show CLI version
-  help: flags.help({ char: 'h' }),
 }
 
 module.exports = GenEnvTemplateCommand
